pyproject.toml
[tool.poetry]
name = "md2gost"
version = "0.5.3"
description = "Скрипт для генерации docx отчета в соответствии с ГОСТ из markdown файла"
authors = ["witelokk <witelokk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-docx = "^1.0.0"
marko = "^2.0.0"
freetype-py = "^2.4.0"
pillow = "^10.0.0"
matplotlib = "^3.7.2"
requests = "^2.31.0"
latex2mathml = "^3.76.0"
pygments = "^2.16.1"
uniseg = "^0.7.2"
fpdf2 = "^2.7.6"
click = "^8.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
md2gost = "md2gost.__main__:main"
